Downloads/Traffic-AI-main/app/requirements.txt
--index-url https://${ARTIFACTORY_USER_ID_ENCODED}:${ARTIFACTORY_ACCESS_TOKEN}@artifact.oocl.com/artifactory/api/pypi/pypi-repository-virtual-external/simple --trusted-host artifact.oocl.com 

rapidfuzz
pyyaml
quart-cors
quart-schema
mongoengine
config-agent==2.9.4
pymongo==4.4.1
typing_extensions==4.12.2
Quart==0.20.0
Hypercorn==0.17.3
ag2==0.7.4
aiofiles==24.1.0
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.8.0
asyncer==0.0.8
autogen-agentchat==0.5.7
autogen-core==0.5.7
autogen-ext[mcp]==0.5.7
azure-ai-inference==1.0.0b8
azure-core==1.32.0
azure-identity==1.20.0
azureml==0.2.7
boto3==1.36.22
botocore==1.36.22
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.1
Deprecated==1.2.18
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
fast-depends==2.4.12
googleapis-common-protos==1.67.0
grpcio==1.70.0
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
isodate==0.7.2
jiter==0.8.2
jmespath==1.0.1
jsonref==1.1.0
msal==1.31.1
msal-extensions==1.2.0
numpy==2.1.3
openai==1.69.0
openlit==1.33.8
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-instrumentation==0.48b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
portalocker==2.10.1
protobuf==5.29.3
pyautogen==0.7.4
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
regex==2024.11.6
requests==2.32.3
s3transfer==0.11.2
schedule==1.2.2
six==1.17.0
sniffio==1.3.1
termcolor==2.5.0
tiktoken==0.7.0
tqdm==4.67.1
tzdata==2025.1
urllib3==2.3.0
websockets==14.2
wrapt==1.17.2
xmltodict==0.13.0
zipp==3.21.0
langfuse
jsonpatch
aiohttp
mlflow==3.1.4
authlib
modelscope==1.15.0